+++ src/android/RESTful.java
@@ -24,6 +24,18 @@ import cz.msebera.android.httpclient.protocol.HTTP;
 
 import java.io.UnsupportedEncodingException;
 
+import java.io.IOException;
+import java.io.OutputStream;
+import java.net.HttpURLConnection;
+import java.net.InetSocketAddress;
+import java.net.NoRouteToHostException;
+import java.net.Socket;
+import java.net.URL;
+import java.net.UnknownHostException;
+
+import java.io.PrintWriter;
+import java.io.StringWriter;
+
 /**
  * This class echoes a string called from JavaScript.
  */
@@ -52,6 +64,16 @@ public class RESTful extends CordovaPlugin {
             String url  = args.getString(2);
             this.checkAdmin(user, pass, url, callbackContext);
             return true;
+        }else if (action.equals("isReachable")) {
+            String url = args.getString(0);
+
+            this.isReachable(url, callbackContext);
+            return true;
+        }else if (action.equals("checkService")) {
+            String url = args.getString(0);
+
+            this.get(url, callbackContext);
+            return true;
         }
         return false;
     }
@@ -71,15 +93,27 @@ public class RESTful extends CordovaPlugin {
                 @Override
                 public void onFailure(int statusCode, Header[] headers, Throwable throwable, JSONObject errorResponse) {
                     // called when response HTTP status is "4XX" (eg. 401, 403, 404)
-                    String errorMsg = "GET, onFailure\n"+
-                    "\nstatusCode: "+Integer.toString(statusCode)+
-                    "\nJSONObject.errorResponse: "+errorResponse.toString()+
-                    "";
+                    // String errorMsg = "GET, onFailure\n"+
+                    // "\nstatusCode: "+Integer.toString(statusCode)+
+                    // "\nJSONObject.errorResponse: "+errorResponse.toString()+
+                    // "";
+
+                    String errorMsg = "onFailure, GET " +
+                        "Status Code: " + Integer.toString(statusCode) + ", " +
+                        "Error message: " + throwable.getMessage() +
+                        " ";
+
                     callbackContext.error(errorMsg);
                 }
 
                 public void onFailure(int statusCode, Header[] headers, String responseString, Throwable throwable) {
-                    callbackContext.error(responseString);
+
+                    String errorMsg = "onFailure, GET " +
+                        "Status Code: " + Integer.toString(statusCode) + ", " +
+                        "Error message: " + throwable.getMessage() +
+                        " ";
+
+                    callbackContext.error(errorMsg);
                 }
 
                 @Override
@@ -123,13 +157,14 @@ public class RESTful extends CordovaPlugin {
         } else {
            callbackContext.error("Get, needs 3 arguments : user: String, pass: String, url: String.");
         }
-    }// post
+    }// get
 
     private void post(String user, String pass, String url, String body, final CallbackContext callbackContext) {
 
         if (user != null && user.length() > 0 || pass != null && pass.length() > 0 || url != null && url.length() > 0 || body != null && body.length() > 0) {
 
             client.setBasicAuth(user, pass);
+            final String urlMsg = url;  
 
             String jsonString   = body;
             JSONArray jsonArray = null;
@@ -169,10 +204,17 @@ public class RESTful extends CordovaPlugin {
                 @Override
                 public void onFailure(int statusCode, Header[] headers, byte[] errorResponse, Throwable e) {
                     // called when response HTTP status is "4XX" (eg. 401, 403, 404)
-                    String errorMsg = "Post, onFailure\n"+
-                    "\nstatusCode: "+Integer.toString(statusCode)+
-                    "\nerrorResponse: "+new String(errorResponse)+
-                    "";
+                    // String errorMsg = "Post, onFailure\n"+
+                    // "\nstatusCode: "+Integer.toString(statusCode)+
+                    // "\nerrorResponse: "+new String(errorResponse)+
+                    // "";
+
+                    String errorMsg = "onFailure, POST " +
+                        "Status Code: " + Integer.toString(statusCode) + ", " +
+                        "Error message: " + e.getMessage() + ", " +
+                        "URL: " + urlMsg +
+                        " ";
+                    
                     callbackContext.error(errorMsg);
                 }
 
@@ -191,7 +233,7 @@ public class RESTful extends CordovaPlugin {
         if (user != null && user.length() > 0 || pass != null && pass.length() > 0 || url != null && url.length() > 0) {
 
             client.setBasicAuth(user, pass);
-            client.setConnectTimeout(2000);
+            //client.setConnectTimeout(2000);
 
             client.get(url, new AsyncHttpResponseHandler() {
 
@@ -210,10 +252,14 @@ public class RESTful extends CordovaPlugin {
                 @Override
                 public void onFailure(int statusCode, Header[] headers, byte[] errorResponse, Throwable e) {
                     // called when response HTTP status is "4XX" (eg. 401, 403, 404)
-                    String errorMsg = "checkAdmin, onFailure\n" +
-                            "\nstatusCode: " + Integer.toString(statusCode) +
-                            "\nJSONObject.errorResponse: " + new String(errorResponse) +
-                            "";
+                    // String errorMsg = "checkAdmin, onFailure\n" +
+                    //         "\nstatusCode: " + Integer.toString(statusCode) +
+                    //         "\nJSONObject.errorResponse: " + new String(errorResponse) +
+                    //         "";
+                    String errorMsg = "onFailure checkAdmin, " +
+                            "Status Code: " + Integer.toString(statusCode) + ", " +
+                            "Error message: " + e.getMessage() +
+                            " ";
     
                     callbackContext.error(errorMsg);
                 }
@@ -229,5 +275,89 @@ public class RESTful extends CordovaPlugin {
             callbackContext.error("CheckAdmin, needs 3 arguments : user: String, pass: String, url: String.");
         }
     }// checkAdmin
+
+
+    private void get(String url, final CallbackContext callbackContext) {
+        final String urlFinal = url;
+        if (url != null && url.length() > 0) {
+
+            client.get(url, new AsyncHttpResponseHandler() {
+
+                @Override
+                public void onStart() {
+                    // called before request is started
+                }
+
+                @Override
+                public void onSuccess(int statusCode, Header[] headers, byte[] response) {
+                    // called when response HTTP status is "200 OK"
+                    String resp = new String(response);
+                    callbackContext.success(resp);
+                }
+
+                @Override
+                public void onFailure(int statusCode, Header[] headers, byte[] errorResponse, Throwable e) {
+                    // called when response HTTP status is "4XX" (eg. 401, 403, 404)
+
+                    // printStackTrace to String
+                    // StringWriter sw = new StringWriter();
+                    // PrintWriter pw = new PrintWriter(sw);
+                    // e.printStackTrace(pw);
+                    // String = printStackTraceString = sw.toString();
+
+                    String errorMsg = "onFailure, " +
+                            "Status Code: " + Integer.toString(statusCode) + ", " +
+                            "URL: " + urlFinal + ", " +
+                            "Error message: " + e.getMessage() +
+                            " ";
+    
+                    callbackContext.error(errorMsg);
+                }
+
+                @Override
+                public void onRetry(int retryNo) {
+                    // called when request is retried
+                }
+            });
+
+
+        } else {
+            callbackContext.error("get simple, needs 1 argument : url: String.");
+        }
+    }// get simple without basic authentication
+
+
+    public boolean isReachable(String targetUrl, final CallbackContext callbackContext)
+    {
+        
+        if (targetUrl != null && targetUrl.length() > 0) {
+        try
+        {
+        	HttpURLConnection httpUrlConnection = (HttpURLConnection) new URL(
+                    targetUrl).openConnection();
+            httpUrlConnection.setRequestMethod("HEAD");
+            int responseCode = httpUrlConnection.getResponseCode();
+            System.out.println("ok");
+            callbackContext.success("ok : "+targetUrl);
+            return responseCode == HttpURLConnection.HTTP_OK;
+
+        } catch (UnknownHostException noInternetConnection)
+        {
+
+        	noInternetConnection.printStackTrace();
+            callbackContext.error("Error not reachable");
+            return false;
+        }
+        catch (Exception e)
+        {	
+        	e.printStackTrace();
+            callbackContext.error("Error not reachable");
+            return false;
+        }
+        }else{
+            callbackContext.error("target url must not be null or empty");
+            return false;
+        }
+    }
 
 }