+++ www/RESTful.js
@@ -12,4 +12,12 @@ exports.checkAdmin = function(arg0, arg1, arg2, success, error) {
     exec(success, error, "RESTful", "checkAdmin", [arg0, arg1, arg2]);
 };
 
+exports.isReachable = function(arg0, success, error) {
+    exec(success, error, "RESTful", "isReachable", [arg0]);
+};
+
+exports.checkService = function(arg0, success, error) {
+    exec(success, error, "RESTful", "checkService", [arg0]);
+};
+