+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## ChangeLog
-#### Version 0.0.3 (12.06.2017)
-
+#### Version 0.0.3 (18.07.2017)
+- Add Get Without Basic Authentication
+- Simplified messages error
 #### Version 0.0.2 
 - Support platform Android